p4i integration Integrate 170609 from 4.0 to main: Fixes:    Use uchar for QLatin1Char::unicode() Task:     none RevBy:    Lars AutoTest: none Details: Use uchar when converting the char to unicode, this ensure we get the entire set of latin1 characters.
[git-p4: depot-paths = "//depot/qt/main/": change = 170629]

+++ src/corelib/tools/qchar.h
@@ -23,7 +23,7 @@ struct QLatin1Char
 public:
     inline explicit QLatin1Char(char c) : ch(c) {}
     inline const char toLatin1() const { return ch; }
-    inline const ushort unicode() const { return ushort(ch); }
+    inline const ushort unicode() const { return ushort(uchar(ch)); }
 
 private:
     const char ch;
